Move tlsconfig to client package.
